Update index.html
I understand that the point of this article is to showcase the simplicity on Surge vs GH. And you've definitely convinced me to try it out. I think the message would carry more value without the unnecessary inflation of GH-Pages steps and not including credentials for Surge.

Even with the proposed changes, the article still shows how much more of a setup there is for git. Even though you aren't writing about the installation of node!
(download, click next a bunch, etc); while at the same time writing about absolutely all of Git's setup (although this is 100% independent of the actual gh-pages publishing in my opinion).